[all] Prevent attempt to clear temporary field data during migration operations

+++ doodle/apps.py
@@ -1,14 +1,18 @@
 from django.apps import AppConfig
 
+import sys
+
 
 class DoodleConfig(AppConfig):
     name = 'doodle'
 
     def ready(self):
-        from . import models
+        ignore_args = ['makemigrations', 'migrate']
+        if not [x for x in ignore_args if x in sys.argv]:
+            from . import models
 
-        # Clear "temporary" fields on room as all websockets were closed (server offline)
-        for room in models.Room.objects.all():
-            room.users.clear()
+            # Clear "temporary" fields on room as all websockets were closed (server offline)
+            for room in models.Room.objects.all():
+                room.users.clear()
 
-        models.Game.objects.all().delete()
+            models.Game.objects.all().delete()
